updates the customer's firstname, lastname, email and country's code on `hookPayment`
Rationale: these fields are only set on the module's `hookHeader`, and that works as long as there's a registered customer at that point. However on one-page-checkout scenarios, the customer might not be set at that point.
Typically, this customer will already be set once `hookPayment` is called, hence we can update the name, email and country variables on the main script, so that they're actually sent with the intent confirmation.